# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: preferences
    description: User UI preferences
  - name: history
    description: Simulation run history
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /preferences/{userId}:
    get:
      operationId: getUserPreferences
      tags: [preferences]
      summary: Get user preferences
      description: Returns stored UI preferences for a given user. Returns defaults when the user has no saved preferences.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            maxLength: 128
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[a-zA-Z0-9]+)$"
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    put:
      operationId: setUserPreferences
      tags: [preferences]
      summary: Save user preferences
      description: Upserts UI preferences for a given user.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            maxLength: 128
            pattern: "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}|[a-zA-Z0-9]+)$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPreferences"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
  /history:
    get:
      operationId: getSimulationHistory
      tags: [history]
      summary: Get simulation run history
      description: Returns the authenticated user's past simulation runs, newest first.
      responses:
        "200":
          description: List of simulation runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SimulationRunSummary"
        "401":
          description: Unauthorized
    post:
      operationId: saveSimulationRun
      tags: [history]
      summary: Save a simulation run
      description: Persists a completed simulation run for the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationRunPayload"
      responses:
        "201":
          description: Run saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationRunSummary"
        "401":
          description: Unauthorized
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    UserPreferences:
      type: object
      additionalProperties: false
      properties:
        minimapVisible:
          type: boolean
        tagsVisible:
          type: boolean
      required:
        - minimapVisible
        - tagsVisible
    ScoreBucketSummary:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        max:
          type: integer
        awarded:
          type: integer
        arithmetic:
          type: string
        reasons:
          type: array
          items:
            type: string
      required:
        - id
        - label
        - max
        - awarded
        - arithmetic
        - reasons
    SimulationScoreData:
      type: object
      properties:
        total:
          type: integer
        buckets:
          type: array
          items:
            $ref: "#/components/schemas/ScoreBucketSummary"
        strengths:
          type: array
          items:
            type: string
        misses:
          type: array
          items:
            type: string
        teachingPoints:
          type: array
          items:
            type: string
      required:
        - total
        - buckets
        - strengths
        - misses
        - teachingPoints
    SimulationRunPayload:
      type: object
      properties:
        scenario:
          type: string
          maxLength: 256
        seed:
          type: string
          maxLength: 128
          nullable: true
        outcome:
          type: string
          maxLength: 64
          nullable: true
        score:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        durationSeconds:
          type: integer
          minimum: 0
          nullable: true
        scoreData:
          $ref: "#/components/schemas/SimulationScoreData"
          nullable: true
      required:
        - scenario
    SimulationRunSummary:
      type: object
      properties:
        id:
          type: string
        scenario:
          type: string
        seed:
          type: string
          nullable: true
        outcome:
          type: string
          nullable: true
        score:
          type: integer
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        scoreData:
          $ref: "#/components/schemas/SimulationScoreData"
          nullable: true
        runAt:
          type: string
          format: date-time
      required:
        - id
        - scenario
        - runAt