# lib/api-zod/src/generated/types/scoreBucketSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreBucketSummary {
  id: string;
  label: string;
  max: number;
  awarded: number;
  arithmetic: string;
  reasons: string[];
}