# lib/api-zod/src/generated/types/simulationRunPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationScoreData } from "./simulationScoreData";

export interface SimulationRunPayload {
  /** @maxLength 256 */
  scenario: string;
  /** @maxLength 128 */
  seed?: string | null;
  /** @maxLength 64 */
  outcome?: string | null;
  /**
   * @minimum 0
   * @maximum 100
   */
  score?: number | null;
  /** @minimum 0 */
  durationSeconds?: number | null;
  scoreData?: SimulationScoreData | null;
}