# lib/api-zod/src/generated/types/simulationRunSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SimulationScoreData } from "./simulationScoreData";

export interface SimulationRunSummary {
  id: string;
  scenario: string;
  seed?: string | null;
  outcome?: string | null;
  score?: number | null;
  durationSeconds?: number | null;
  scoreData?: SimulationScoreData | null;
  runAt: Date;
}