# lib/api-zod/src/generated/types/simulationScoreData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreBucketSummary } from "./scoreBucketSummary";

export interface SimulationScoreData {
  total: number;
  buckets: ScoreBucketSummary[];
  strengths: string[];
  misses: string[];
  teachingPoints: string[];
}